rookie-security-encrypt/src/main/java/com/rookie/bigdata/basic/nationalpassword/sm3/SM3Util.java
package com.rookie.bigdata.basic.nationalpassword.sm3;

import org.bouncycastle.crypto.digests.SM3Digest;
import org.bouncycastle.crypto.macs.HMac;
import org.bouncycastle.crypto.params.KeyParameter;
import org.bouncycastle.jce.provider.BouncyCastleProvider;


import java.security.Security;
import java.util.Arrays;

/**
 * @Class SM3Util
 * @Description
 * @Author rookie
 * @Date 2024/7/23 18:04
 * @Version 1.0
 */
public class SM3Util /*extends BaseUtil*/ {



    static {
        Security.addProvider(new BouncyCastleProvider());
    }

    /**
     *
     * @param srcData
     * @return
     */
    public static byte[] hash(byte[] srcData) {
        SM3Digest digest = new SM3Digest();
        digest.update(srcData, 0, srcData.length);
        byte[] hash = new byte[digest.getDigestSize()];
        digest.doFinal(hash, 0);
        return hash;
    }

    /**
     *
     * @param key
     * @param srcData
     * @return
     */
    public static byte[] hmac(byte[] key, byte[] srcData) {
        KeyParameter keyParameter = new KeyParameter(key);
        SM3Digest digest = new SM3Digest();
        HMac mac = new HMac(digest);
        mac.init(keyParameter);
        mac.update(srcData, 0, srcData.length);
        byte[] result = new byte[mac.getMacSize()];
        mac.doFinal(result, 0);
        return result;
    }

    /**
     *
     * @param srcData
     * @param sm3Hash
     * @return
     */
    public static boolean verify(byte[] srcData, byte[] sm3Hash) {
        byte[] newHash = hash(srcData);
        if (Arrays.equals(newHash, sm3Hash)) {
            return true;
        } else {
            return false;
        }
    }
}
